Move option services to separate folder
